separate buildProjectMap for reuse in different SET____TIMESTAMPS contexts
