+++ src/payloads.js
@@ -5,5 +5,5 @@ export const FETCH_CONSULTANT_TIMESTAMPS_PAYLOAD = (consultantId, start_time, en
         ^ORDERBYstart_time`
         :
         `consultant=${consultantId}^start_timeONToday@javascript:gs.beginningOfToday()@javascript:gs.endOfToday()^ORDERBYstart_time`,
-    sysparm_fields: `project.client.short_description, project.sys_id, project.short_description, start_time,end_time,active,duration,rounded_duration,sys_id,note`
+    sysparm_fields: `project.client.short_description, project.client.sys_id, project.sys_id, project.short_description, start_time,end_time,active,duration,rounded_duration,sys_id,note`
 })
