+++ src/x-esg-timer-container/actionHandlers.js
@@ -1,6 +1,6 @@
 import { actionTypes } from '@servicenow/ui-core';
 import { createHttpEffect } from '@servicenow/ui-effect-http';
-import { getSnDayBounds} from '../helpers';
+import { buildProjectMap, getSnDayBounds} from '../helpers';
 import { FETCH_CONSULTANT_TIMESTAMPS_PAYLOAD } from '../payloads';
 
 const {COMPONENT_BOOTSTRAPPED} = actionTypes;
@@ -82,40 +82,6 @@ export default {
         errorActionType: 'LOG_ERROR',
     }),
     'SET_CONSULTANT_TIMESTAMPS': ({action, updateState}) => {
-        console.log('Setting timestamps: action:', action);
-        const timestamps = action.payload.result;
-        const stampsByProject = new Map();
-        
-        // Massage for easy mapping
-        // Subtracting the parsed ServiceNow zero duration time with 
-        // Date.parse("1970-01-01 00:00:00") corrects for timezone issues, etc.
-        for(let stamp of timestamps){
-            const projectId = stamp['project.sys_id'];
-            const active = stamp.active === 'true';
-
-            const sharedValues = {
-                active,
-                sys_id: projectId,
-                note: stamp.note,
-                client: stamp['project.client.short_description'],
-                short_description: stamp['project.short_description'],
-            }
-            if(stampsByProject.has(projectId)){
-                stampsByProject.set(projectId, {
-                    ...sharedValues,
-                    timestamps: [stamp, ...stampsByProject.get(projectId).timestamps],
-                    totalRoundedTime: stampsByProject.get(projectId).totalRoundedTime + 
-                        (Date.parse(stamp.rounded_duration) - Date.parse("1970-01-01 00:00:00") 
-                        || 0),
-                });
-            } else{
-                stampsByProject.set(projectId, {
-                    ...sharedValues,
-                    timestamps: [stamp],
-                    totalRoundedTime: Date.parse(stamp.rounded_duration) - Date.parse("1970-01-01 00:00:00") || 0,
-                })
-            }
-        }
-        updateState({projectMap: stampsByProject});
+        updateState({projectMap: buildProjectMap(action.payload.result)});
     },
 }
