+++ src/x-esg-timesheet-app/index.js
@@ -13,7 +13,7 @@ createCustomElement('x-esg-timesheet-app', {
 	styles,
 	initialState: {
 		consultantId: '',
-		location: 'week',
+		location: 'day',
 		addProjectStatus: false,
 		editMode: false,
 		genericProjects: [],
